bugs

+++ Hangman/HighScore Form.cs
@@ -18,40 +18,39 @@ namespace Hangman
         public Form2()
         {
             InitializeComponent();
-            encDecr en = new encDecr();
-            en.DecryptFile(outp, decr);
-            File.Delete(outp);
-            StreamReader read = new StreamReader(decr);
-            String line;
-            int c = 0;
-            int i = 0;
-            Label[] labels = new Label[] { label2, label3, label4, label5, label6, label7, label8, label9, label10, label11 };
-            while (c < 5 && (line = read.ReadLine()) != null)
+            if (File.Exists(outp))
             {
-                var d = line.Split(' ');
-                String name = d[0];
-                //int score = int.Parse(d[1]);
-                String score = d[1];
-                labels[i].Text = name;
-                i++;
-                labels[i].Text = score;
-                i++;
-                c++;
+                encDecr en = new encDecr();
+                en.DecryptFile(outp, decr);
+                File.Delete(outp);
+                StreamReader read = new StreamReader(decr);
+                String line;
+                int c = 0;
+                int i = 0;
+                Label[] labels = new Label[] { label2, label3, label4, label5, label6, label7, label8, label9, label10, label11 };
+                while (c < 5 && (line = read.ReadLine()) != null)
+                {
+                    var d = line.Split(' ');
+                    String name = d[0];
+                    //int score = int.Parse(d[1]);
+                    String score = d[1];
+                    labels[i].Text = name;
+                    labels[i].Visible = true;
+                    i++;
+                    labels[i].Text = score;
+                    labels[i].Visible = true;
+                    i++;
+                    c++;
 
+                }
+                read.Close();
+                en.EncryptFile(decr, outp);
+                File.Delete(decr);
             }
-            read.Close();
-            en.EncryptFile(decr, outp);
-            File.Delete(decr);
         }
         public void OnBackToMainClicked(object sender, EventArgs e)
         {
             Application.Exit();
-        }
-        
-        private void fsdf()
-        {
-            /*encDecr en = new encDecr();
-            en.DecryptFile(outp,decr);*/
         }
 
         private void Form2_FormClosed(object sender, FormClosedEventArgs e)
